requirements.txt
# ── Web Framework ─────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0

# ── Database ───────────────────────────────────────────────
neo4j==5.19.0

# ── Auth & Security ────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# ── Data validation ────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# ── Environment ────────────────────────────────────────────
python-dotenv==1.0.1

# ── AI / ML (ready for future predictions) ─────────────────
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
joblib==1.4.2

# ── Utilities ──────────────────────────────────────────────
python-dateutil==2.9.0
